fix: move slot.tsx to ui/components/ui/ to prevent duplicate barefoot.js loading
Move ui/base/slot.tsx to ui/components/ui/slot.tsx so all components
share the same directory structure. This ensures relative imports for
barefoot.js resolve to the same path, preventing the runtime from being
loaded twice from different URLs (/static/barefoot.js and
/static/components/barefoot.js).

Changes:
- Move ui/base/slot.tsx to ui/components/ui/slot.tsx
- Update button.tsx to import Slot from ./slot
- Remove base/ directory handling from docs/ui/build.ts
- Remove duplicate barefoot.js copy to dist/ root

Closes #191
